fix: lint and naming issues
